+++ src/ldpcMinSum.cpp
@@ -13,7 +13,7 @@ ldpcMinSumDec::ldpcMinSumDec(ldpcMatrix const &mat) :
   scale = 0.75;
 }
 
-std::vector<int> ldpcMinSumDec::decode(std::vector<double> const &llr)
+std::vector<int> ldpcMinSumDec::decode(std::vector<float> const &llr)
 {
   hd.assign(hd.size(), 1);
   llrIn = llr;
@@ -30,7 +30,7 @@ std::vector<int> ldpcMinSumDec::decode(std::vector<double> const &llr)
 
 int ldpcMinSumDec::update_v2c()
 {
-  std::vector<double> sum_llr = llrIn;
+  std::vector<float> sum_llr = llrIn;
 
   int ind = 0;
   for(auto it = g.begin(); it != g.end(); ++it, ++ind) {
@@ -41,7 +41,7 @@ int ldpcMinSumDec::update_v2c()
     v2c[ind] = sum_llr[it->col] - c2v[ind];
   }
 
-  std::transform(sum_llr.begin(), sum_llr.end(), hd.begin(), std::bind2nd(std::less<double>(), 0.0));
+  std::transform(sum_llr.begin(), sum_llr.end(), hd.begin(), std::bind2nd(std::less<float>(), 0.0));
 
   return 0;
 }
@@ -49,12 +49,12 @@ int ldpcMinSumDec::update_v2c()
 int ldpcMinSumDec::update_c2v()
 {
   std::vector<int> sign(numRows, 1);
-  std::vector<double> min1(numRows, 1e32), min2(numRows, 1e32);
+  std::vector<float> min1(numRows, 1e32), min2(numRows, 1e32);
   std::vector<int> min_ind(numRows, -1);
 
   int ind = 0;
   for(auto it = g.begin(); it != g.end(); ++it, ++ind) {
-    double mag = fabs(v2c[ind]);
+    float mag = fabs(v2c[ind]);
     sign[it->row] *= (v2c[ind] < 0) ? -1 : 1;
     if(mag < min1[it->row]) {
       min2[it->row] = min1[it->row];
